+++ CHANGELOG.md
@@ -15,6 +15,7 @@ We haven't had the best track record of code/feature changes before this date, b
 - [fbt-rn-android-native versions](#fbt-rn-android-native)
 
 ### Top-level Github
+- Remove the String.prototype methods (that aren't used internally) from libdefs
 - Upgrade to Flow v0.113.0
 - Make demo-app rely on the local versions of our published npm modules.
   That'll avoid us from having to update `package.json` for every package update.